# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Diagnostic Center Billing ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientList"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    put:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/history:
    get:
      operationId: getPatientHistory
      tags: [patients]
      summary: Get patient order history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"

  # Tests
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List diagnostic tests
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tests list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestList"
    post:
      operationId: createTest
      tags: [tests]
      summary: Create diagnostic test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestBody"
      responses:
        "201":
          description: Test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticTest"
  /tests/{id}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get test by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticTest"
    put:
      operationId: updateTest
      tags: [tests]
      summary: Update test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestBody"
      responses:
        "200":
          description: Updated test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticTest"

  # Orders
...
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        bloodGroup:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, patientId, firstName, lastName, dateOfBirth, gender, phone, createdAt]
    PatientList:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [patients, total, page, limit]
    CreatePatientBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        bloodGroup:
          type: string
          nullable: true
      required: [firstName, lastName, dateOfBirth, gender, phone]
    DiagnosticTest:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        duration:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, code, name, category, price, duration, isActive, createdAt]
    TestList:
      type: object
      properties:
        tests:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticTest"
        total:
          type: integer
      required: [tests, total]
    CreateTestBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        duration:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [code, name, category, price, duration]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        patientId:
          type: integer
        patient:
          $ref: "#/components/schemas/Patient"
        doctorId:
          type: integer
          nullable: true
        doctor:
          $ref: "#/components/schemas/Doctor"
          nullable: true
        status:
          type: string
          enum: [pending, collected, processing, completed, cancelled]
        tests:
          type: array
          items:
            $ref: "#/components/schemas/OrderTest"
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        collectedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderNumber, patientId, patient, status, tests, totalAmount, createdAt]
    OrderTest:
      type: object
      properties:
        id:
          type: integer
        testId:
          type: integer
        test:
          $ref: "#/components/schemas/DiagnosticTest"
        price:
          type: number
        result:
          type: string
          nullable: true
        resultStatus:
          type: string
          nullable: true
      required: [id, testId, test, price]
    OrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [orders, total, page, limit]
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        phone:
          type: string
          nullable: true
        hospitalAffiliation:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, specialization, createdAt]
...